pyproject.toml
[project]
name = "scraping-uber"
version = "0.1.0"
description = "Production-grade web scraping pipeline for food delivery platforms (UberEats, DoorDash, Grubhub, JustEatTakeaway)"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = []
keywords = ["scraping", "food-delivery", "airflow", "playwright"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "playwright>=1.44",
    "scrapy>=2.11",
    "scrapy-playwright>=0.0.40",
    "beautifulsoup4>=4.12",
    "selectolax>=0.3",
    "httpx>=0.27",
    "pydantic>=2.7",
    "sqlalchemy>=1.4,<2.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.13",
    "geoalchemy2>=0.15",
    "boto3>=1.34",
    "minio>=7.2",
    "splink>=4.0",
    "rapidfuzz>=3.9",
    "duckdb>=1.0",
    "confluent-kafka>=2.4.0",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "prometheus-client>=0.20",
    "geopy>=2.4",
    "haversine>=2.8",
    "pandas>=2.2",
    "pyarrow>=16.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "slack-sdk>=3.31",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["scrapers*", "processing*", "storage*", "streaming*", "analytics*"]
exclude = ["*tests*", "*examples*", "*scripts*", "local_storage*", "dags*", "config*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.pyright]
extraPaths = ["."]

[tool.ruff]
target-version = "py39"
line-length = 100
exclude = [".git", ".venv", "__pycache__", "build", "dist", "local_storage"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
